finish

+++ application/controllers/Posts.php
@@ -1,9 +1,19 @@
 <?php
 	class Posts extends CI_Controller
 	{
-		public function index(){
+		public function index($offset = 0){
+			//pagination config
+			$config['base_url'] = base_url().'posts/index/';
+			$config['total_rows'] = $this->db->count_all('posts');
+			$config['per_page'] = 3;
+			$config['uri_segment'] = 3;
+			$config['attributes'] = array('class' => 'pagination-link');
+
+			//init pagination
+			$this->pagination->initialize($config);
+
 			$data['title'] = 'Latest Posts';
-			$data['posts'] = $this->post_model->get_posts();
+			$data['posts'] = $this->post_model->get_posts(FALSE, $config['per_page'], $offset);
  
 			$this->load->view('templates/header');
 			$this->load->view('posts/index', $data);
@@ -25,6 +35,12 @@
 		}
 
 		public function create(){
+			//check login
+			if(!$this->session->userdata('logged_in')){
+				$this->session->set_flashdata('user_notlogin', 'You must login first!');
+				redirect('users/login');
+			}
+
 			$data['title'] = 'Create Post';
 			$data['categories'] = $this->post_model->get_categories();
 
@@ -32,6 +48,7 @@
 			$this->form_validation->set_rules('body', 'Body', 'required');
 
 			if(!$this->form_validation->run()){
+
 			$this->load->view('templates/header');
             $this->load->view('posts/create', $data); 
             $this->load->view('templates/footer'); 
@@ -58,17 +75,44 @@
 				}
 
 				$this->post_model->create_post($post_image);
+				
+				//set message
+				$this->session->set_flashdata('post_created', 'Your post has been created!');
+
+
 				header("location: ".base_url()."posts");
 			}
 		}
 
 		public function delete($id){
+			//check if logged in 
+			if(!$this->session->userdata('logged_in')){
+				$this->session->set_flashdata('user_notlogin', 'You must login first!');
+				redirect('users/login');
+			}
+			if($this->session->userdata('user_id') != $id){
+				redirect('posts');
+			}
 			$this->post_model->delete_post($id);
+			//set message
+			$this->session->set_flashdata('post_deleted', 'Post Has been Deleted!');
+
 			header("location: ".base_url()."posts");
 		}
 
 		public function edit($slug){
+			if(!$this->session->userdata('logged_in')){
+				$this->session->set_flashdata('user_notlogin', 'You must login first!');
+				redirect('users/login');
+			}
+
 			$data['post'] = $this->post_model->get_posts($slug);
+
+			//check user
+			if($this->session->userdata('user_id') != $this->post_model->get_posts($slug)['user_id']){
+				redirect('posts');
+			}
+		
 			$data['categories'] = $this->post_model->get_categories();
 
 
@@ -84,7 +128,16 @@
 		}
 
 		public function update(){
+			if(!$this->session->userdata('logged_in')){
+				$this->session->set_flashdata('user_notlogin', 'You must login first!');
+				redirect('users/login');
+			}
+
 			$this->post_model->update_post();
+
+			//set message
+			$this->session->set_flashdata('post_updated', 'Post Has been updated!');
+			
 			redirect('posts');
 		}
 	}
